# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API cho Superapp dành cho Tài xế công nghệ.\nBao gồm: Auth, GPS Tracking, SOS, Chat, Map Sinh tồn, Mood, Earning, Reminder, Emergency Contact.",
        "title": "Grab Driver Superapp API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reminder/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mobile gọi để kiểm tra đã đến lúc nhắc nước/vận động chưa. Nếu key Redis hết TTL → nhắc + reset timer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reminder"
                ],
                "summary": "Kiểm tra trạng thái nhắc nhở",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_reminder.ReminderStatus"
                        }
                    }
                }
            }
        },
        "/reminder/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy cấu hình nhắc uống nước, vận động. Tạo mặc định nếu chưa có.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reminder"
                ],
                "summary": "Lấy cấu hình nhắc nhở",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_reminder.ReminderConfig"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bật/tắt nhắc nước, vận động, thay đổi interval, giờ yên tĩnh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reminder"
                ],
                "summary": "Cập nhật cấu hình nhắc nhở",
                "parameters": [
                    {
                        "description": "Cấu hình mới",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_reminder.UpdateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_reminder.ReminderConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_modules_reminder.ReminderConfig": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "quiet_end": {
                    "type": "string"
                },
                "quiet_start": {
                    "description": "HH:MM — không nhắc ban đêm",
                    "type": "string"
                },
                "stretch_enabled": {
                    "type": "boolean"
                },
                "stretch_interval_min": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "water_enabled": {
                    "type": "boolean"
                },
                "water_interval_min": {
                    "description": "phút",
                    "type": "integer"
                }
            }
        },
        "internal_modules_reminder.ReminderStatus": {
            "type": "object",
            "properties": {
                "driver_id": {
                    "type": "integer"
                },
                "next_stretch_at": {
                    "type": "integer"
                },
                "next_water_at": {
                    "description": "unix timestamp",
                    "type": "integer"
                }
            }
        },
        "internal_modules_reminder.UpdateConfigRequest": {
            "type": "object",
            "properties": {
                "quiet_end": {
                    "type": "string"
                },
                "quiet_start": {
                    "type": "string"
                },
                "stretch_enabled": {
                    "type": "boolean"
                },
                "stretch_interval_min": {
                    "type": "integer"
                },
                "water_enabled": {
                    "type": "boolean"
                },
                "water_interval_min": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập JWT token theo format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}